Add form logic
This adds the Form component which takes a numDigits prop and an
onSubmit prop.

The form displays a number of boxes and updates the focus to the next
input as the current input has input given to it.

The the last box is given input and all of the other inputs have
been filled, the form calls onSubmit automatically with the code
input across all of the inputs.